+++ lib/dart_rip_log.dart
@@ -32,3 +32,5 @@ export 'src/parser.dart' show parseRipLog, detectLogFormat;
 export 'src/file_reader.dart' show parseRipLogFile;
 export 'src/convenience.dart'
     show isFullyVerified, tracksWithErrors, tracksWithArMismatch, toJson;
+export 'src/diff.dart'
+    show compareRipLogs, RipLogDiff, RipLogDiffEntry, RipLogDiffKind;